pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "pybind11>=2.10"]
build-backend = "setuptools.build_meta"

[project]
name = "src-cascades"
version = "3.0.0"
authors = [ { name = "Anton L. Waagaard" } ]
description = "A library for simulating, analyzing, and fitting self-reinforcing cascade models to data."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.21",
    "networkx>=2.6",
    "pandas>=1.3",
    "matplotlib>=3.5",
    "sympy>=1.9",
    "pyyaml>=6.0",
    "scipy>=1.8",
    "emcee>=3.1",
    "arviz>=0.12",
]

[project.urls]
Homepage = "https://github.com/anton-lw/src-cascades-project"

[tool.setuptools]
py-modules = ["src_cascades.simulator_backend_cpp"]

[tool.setuptools.packages.find]
where = ["src"]